Move domain validation checks for URI/Username to service startup
These checks were done once per request. These instead can be static
checks done on service startup. This enforces that the Username and URI
configs appropriately set the SubjectDomain and IssuerUrl fields, with
no scheme just for the SubjectDomain field for Username.
